pyproject.toml
[tool.poetry]
name = "aboba"
version = "0.1.0"
description = ""
authors = ["TeranozavrSex <bananos203@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.4"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.6.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^23.0.0"
drf-spectacular = "^0.28.0"
django-defender = "^0.9.8"
django-redis = "^5.4.0"
validate-email = "^1.3"
argon2-cffi = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
